integrate-with-getprojects
