+++ tests/api/Jalasoft.TeamUp.Projects.API.Tests/GetProjectsTest.cs
@@ -3,6 +3,7 @@
     using Jalasoft.TeamUp.Projects.API.Controllers;
     using Jalasoft.TeamUp.Projects.Core.Interfaces;
     using Jalasoft.TeamUp.Projects.Models;
+    using Jalasoft.TeamUp.Projects.ProjectsException;
     using Microsoft.AspNetCore.Http;
     using Microsoft.AspNetCore.Mvc;
     using Moq;
@@ -26,6 +27,21 @@
         {
             // Arrange
             var request = this.mockHttpContext.Request;
+
+            // Act
+            var response = this.getProjects.Run(request);
+
+            // Assert
+            var okObjectResult = Assert.IsType<OkObjectResult>(response);
+            Assert.IsType<Project[]>(okObjectResult.Value);
+        }
+
+        [Fact]
+        public void GetProjectsBySkill_Returns_OkObjectResult()
+        {
+            // Arrange
+            var request = this.mockHttpContext.Request;
+            this.mockService.Setup(service => service.GetProjectsBySkill("C#")).Returns(new Project[10]);
 
             // Act
             var response = this.getProjects.Run(request);
